api tests work great
