+++ test/api/jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  globals: {
+    "ts-jest": {
+      babelConfig: true
+    }
+  },
+  setupFilesAfterEnv: ["./setup/after-env.ts"],
+  maxWorkers: 1
+}
