+++ test/shared/data/count-by-typeof.zson
