+++ test/shared/data/custom-sample.ndjson
