+++ test/shared/data/index.ts
@@ -1,0 +1,9 @@
+import path from "path"
+
+class Data {
+  getPath(name) {
+    return path.join(__dirname, name)
+  }
+}
+
+export default new Data()
